refactor: 줄 생성 인터페이스 제거 및 연결 상태 전략 적용으로 설계 개선
